Fixed an issue where the frame rate couldn't print after time revision.

+++ src/com/base/engine/core/CoreEngine.java
@@ -66,7 +66,7 @@ public class CoreEngine
 		isRunning = true;
 		
 		int frames = 0;
-		long frameCounter = 0;
+		double frameCounter = 0;
 
 		game.init();
